# src/main/resources/DocumentationOpenAPI.yml
openapi: 3.0.3
info:
  title: Simple Product API
  description: A simple API for managing products.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local server
paths:
  /products:
    get:
      summary: Get all products
      operationId: getProducts
      responses:
        '200':
          description: A list of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    post:
      summary: Create a new product
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Product created
        '400':
          description: Product didnt creat. Bad Request
  /products/{id}:
    get:
      summary: Get a product by ID
      operationId: getProductById
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the product
          schema:
            type: integer
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Product not found. Bed request
        '404':
          description: Product not found
    put:
      summary: Update the product
      operationId: createProduct
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the product
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Product didnt update. Bad Request
        '404':
          description: Product not found by ID
    delete:
      summary: Delete a product by ID
      operationId: deleteProductById
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the product
          schema:
            type: integer
      responses:
        '200':
          description: Product deleted
        '404':
          description: Product not found
  /products/category/{category}:
    get:
      summary: Get a product by category
      operationId: getProductByCategory
      parameters:
        - name: category
          in: path
          required: true
          description: The category of the product
          schema:
            type: string
      responses:
        '200':
          description: A list of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: integer
          description: The product ID
        name:
          type: string
          description: The name of the product
        description:
          type: string
          description: The description of the product
        price:
          type: number
          format: BigDecimal
          description: The price of the product
        category:
          type: string
          description: The category of the product
        stock:
          type: integer
          description: The available stock
        created:
          type: string
          format: date-time
          description: The creation date
        lastUpdated:
          type: string
          format: date-time
          description: The last update date